requirements.txt
Pillow>=10.0
segno>=1.5
